Moved logic to DoctrineProxySubscriber
