+++ tests/JMS/Serializer/Tests/Serializer/EventDispatcher/Subscriber/DoctrineProxySubscriberTest.php
@@ -18,6 +18,7 @@
 
 namespace JMS\Serializer\Tests\Serializer\EventDispatcher\Subscriber;
 
+use JMS\Serializer\EventDispatcher\EventDispatcher;
 use JMS\Serializer\EventDispatcher\PreSerializeEvent;
 use JMS\Serializer\EventDispatcher\Subscriber\DoctrineProxySubscriber;
 use JMS\Serializer\Tests\Fixtures\SimpleObjectProxy;
@@ -30,6 +31,11 @@ class DoctrineProxySubscriberTest extends \PHPUnit_Framework_TestCase
 
     /** @var DoctrineProxySubscriber */
     private $subscriber;
+
+    /**
+     * @var EventDispatcher
+     */
+    private $dispatcher;
 
     public function testRewritesProxyClassName()
     {
@@ -60,10 +66,43 @@ class DoctrineProxySubscriberTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($obj->__isInitialized());
     }
 
+    public function testEventTriggeredOnRealClassName()
+    {
+        $proxy = new SimpleObjectProxy('foo', 'bar');
+
+        $realClassEventTriggered1 = false;
+        $this->dispatcher->addListener('serializer.pre_serialize', function () use (&$realClassEventTriggered1) {
+            $realClassEventTriggered1 = true;
+        }, get_parent_class($proxy));
+
+        $event = $this->createEvent($proxy, array('name' => get_class($proxy), 'params' => array()));
+        $this->dispatcher->dispatch('serializer.pre_serialize', get_class($proxy), 'json', $event);
+
+        $this->assertTrue($realClassEventTriggered1);
+    }
+
+    public function testListenersCanChangeType()
+    {
+        $proxy = new SimpleObjectProxy('foo', 'bar');
+
+        $realClassEventTriggered1 = false;
+        $this->dispatcher->addListener('serializer.pre_serialize', function (PreSerializeEvent $event) use (&$realClassEventTriggered1) {
+            $event->setType('foo', ['bar']);
+        }, get_parent_class($proxy));
+
+        $event = $this->createEvent($proxy, array('name' => get_class($proxy), 'params' => array()));
+        $this->dispatcher->dispatch('serializer.pre_serialize', get_class($proxy), 'json', $event);
+
+        $this->assertSame(['name' => 'foo', 'params' => ['bar']], $event->getType());
+    }
+
     protected function setUp()
     {
         $this->subscriber = new DoctrineProxySubscriber();
         $this->visitor = $this->getMock('JMS\Serializer\Context');
+
+        $this->dispatcher = new EventDispatcher();
+        $this->dispatcher->addSubscriber($this->subscriber);
     }
 
     private function createEvent($object, array $type)
